% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssi.R
\name{ssi}
\alias{ssi}
\title{Generate null distributions and calculate synanthropy index}
\usage{
ssi(r = raster, data = dataset, resolution = 200, sim = 500, threshold = 30)
}
\arguments{
\item{r}{The raster to be analysed (default name: raster)}

\item{data}{The species dataset to be processed, containing columns named
species, year, X, and Y (default name: dataset)}

\item{resolution}{Resolution(s) to be tested, specified as the aggregation factor,
i.e., the number of raster cells in each direction (horizontally and vertically) (default value: 200)}

\item{sim}{Number of simulations (null distributions) to be generated (default value: 500)}

\item{threshold}{Threshold of species occurrence under which the species will
not be accounted for (default value: 30)}
}
\value{
Three data frames:
\code{speciesScores} a short summary table (one line per species) with the mean of all effect
size and the corresponding index (range between 1 to 10) for each resolution.
The number of runs used to calculate the mean effect size is also specified.
\code{effSizes} a data.frame compiling all the raw results, i.e, all the effect sizes
per run, with corresponding information provided by the function cohens_d
(rstatix package); n1 and n2 correspond to the number of occurrences compared
(n1 for the null distribution and n2 for observed data).
\code{samplesList} a data.frame archiving all the occurrence randomly drawn to assess scores.
}
\description{
Generate null distributions and calculate synanthropy index
}
\examples{
example <- ssi(dataset)
}
